Add status filter to idea overview and ability to map statuses
The status filter is customizable in the statuses it should show and
filter on. Each status is selectable through the idea overview options.

This commit also adds the ability to map statuses to a different status
entirely. This means that you can map for instance BUSY to DENIED.
All ideas in the idea overview with the BUSY status will the have the
DENIED status. This combines with the status filter to allow more
configuration over which ideas to show under which conditions.