Fix budgeting reflow
Adds the @zoom400 and @zoom500 media queries which allow checking if a user is zoomed in 400% / 500%. The sticky header is now styled based upon these media queries to allow the user to see more of the budgeting widget when zoomed in at these percentages.

The updateBudgetDisplay function is now called upon window resize, which allows for a user to zoom in/out without reflow issues.