Adds keyboard accessibility to the trix toolbar
The trix toolbar is not keyboard-accessible (see https://github.com/basecamp/trix/issues/186). The solutions provided in this issue do not work either.

This commit adds a 'keypress' event handler to catch any Enter or Space presses on the trix toolbar buttons to ensure the trix editor is useable with just a keyboard.